nlp-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.35.0
torch==2.1.0
pydantic==2.4.0
pydantic-settings==2.0.0
python-dotenv==1.0.0
datasets==2.14.0
aiofiles==23.2.1
python-json-logger==2.0.7
PyYAML==6.0.1
scikit-learn==1.3.2
seqeval==1.2.2
matplotlib==3.8.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0
